Fix bugs in libft

+++ libft/ft_dmtxclear.c
@@ -6,7 +6,7 @@
 /*   By: amamian <amamian@student.42yerevan.am>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/27 18:49:21 by amamian           #+#    #+#             */
-/*   Updated: 2021/08/17 19:44:05 by amamian          ###   ########.fr       */
+/*   Updated: 2021/08/19 14:06:22 by amamian          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
  */
 void	ft_dmtxclear(t_dmtx *dmtx)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (i < dmtx->len)
